+++ django/applications/catmaid/static/js/WindowMaker.js
@@ -2576,6 +2576,7 @@ var WindowMaker = new function()
             '<th>z</th>' +
             '<th>s</ht>' +
             '<th>confidence</ht>' +
+            '<th>target confidence</ht>' +
             '<th>tags</th>' +
             '<th id="connector_nr_nodes_top' + CT.widgetID + '"># nodes for target(s)</th>' +
             '<th>username</th>' +
@@ -2592,6 +2593,7 @@ var WindowMaker = new function()
             '<th>z</th>' +
             '<th>s</ht>' +
             '<th>confidence</ht>' +
+            '<th>target confidence</ht>' +
             '<th>tags</th>' +
             '<th id="connector_nr_nodes_bottom' + CT.widgetID + '"># nodes for target(s)</th>' +
             '<th>username</th>' +